refactor: split utils module into package
