+++ app/openai_batch.py
@@ -7,14 +7,9 @@ import json
 import pandas as pd
 from typing import List, Dict, Any, Optional, Tuple
 from dotenv import load_dotenv
-from .utils import (
-    sanitize_input,
-    write_jsonl,
-    read_jsonl,
-    calculate_batch_size,
-    validate_api_key,
-    estimate_batch_cost,
-)
+from utils.validators import sanitize_input, validate_api_key, calculate_batch_size
+from utils.parsers import write_jsonl, read_jsonl
+from utils.formatters import estimate_batch_cost
 from .logger import logger
 
 from .postprocessor import validate_openai_response
