+++ utils/__init__.py
@@ -1,0 +1,42 @@
+"""Utility package consolidating helper modules."""
+
+from .validators import sanitize_input, validate_api_key, is_valid_poll_interval, calculate_batch_size
+from .formatters import (
+    format_batch_results,
+    calculate_cost_estimate,
+    estimate_batch_cost,
+    format_time_elapsed,
+    generate_unique_id,
+    format_timestamp,
+    calculate_completion_time,
+    export_results_to_csv,
+    format_batch_summary,
+)
+from .parsers import (
+    write_jsonl,
+    read_jsonl,
+    generate_hash,
+    deduplicate_prompts,
+    expand_results,
+)
+
+__all__ = [
+    "sanitize_input",
+    "validate_api_key",
+    "is_valid_poll_interval",
+    "calculate_batch_size",
+    "format_batch_results",
+    "calculate_cost_estimate",
+    "estimate_batch_cost",
+    "format_time_elapsed",
+    "generate_unique_id",
+    "format_timestamp",
+    "calculate_completion_time",
+    "export_results_to_csv",
+    "format_batch_summary",
+    "write_jsonl",
+    "read_jsonl",
+    "generate_hash",
+    "deduplicate_prompts",
+    "expand_results",
+]
